improve feature drawing bank card and risk card
