+++ src/main/java/at/aau/serg/websocketdemoserver/model/board/TileType.java
@@ -8,5 +8,4 @@ public enum TileType {
     PRISON,
     GOTO_JAIL,
     TAX,
-    SPECIAL
 }
