+++ src/main/java/at/aau/serg/websocketdemoserver/model/cards/RiskCardEffect.java
@@ -4,6 +4,5 @@ public enum RiskCardEffect {
     CASH,
     ESCAPE_CARD,
     GOTO_JAIL,
-    TEXT_ONLY
 }
