--- src/main/java/at/aau/serg/websocketdemoserver/service/lobby_request/InitPlayerRequest.java
@@ -1,4 +1,0 @@
-package at.aau.serg.websocketdemoserver.service.lobby_request;
-
-public class InitPlayerRequest {
-}
